requirements.txt
aniso8601==9.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
grpcio==1.35.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
NEURON==7.8.0.119
protobuf==3.15.2
pydgraph==20.7.0
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
